SF : Add support for rotated panel mounts
This is a combination of 3 commits.

SF: Add support for inverse mounted panels

Add 180 rotation in SF to account for inverse mounted panels.
The framework will be unaffected and use policies for a 0 mounted
panel. SF changes the global transform and silently induces the H,V
flips. Similar flips are added to screenshots as well.

Change-Id: I6e9576ee734ee85097491eaa1e8e94cfb3731e0f

SF: Add support for all flips of panel mount

Add support for all flips of panel mount, H, V, HV (180). Property
persist.panel.mountflip can be set to 1 for H-Flip, 2 for V-Flip,
3 for HV-Flip (180 / inverse mount).

Syntax :
    adb shell setprop persist.panel.mountflip 1  -- For H-Flip

Change-Id: Ide7b8378ad6a423e5d7335fedc27d480a25b53ae
CRs-fixed: 990622

SF: Flip the sourceCrop for inverse mounted panel.

Flip the sourceCrop for inverse mounted panel when
screenshot is taken for WindowManagerService

Change-Id: I08bb7e0feea69dd15b372e9fbbd6984c6bfbec46
CRs-Fixed: 1033965

Change-Id: Ide7b8378ad6a423e5d7335fedc27d480a25b53af
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>